+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Dynamic endpoints (WIP)
+
 3.131.5 (2022-07-28)
 ------------------